@@ -4,38 +4,35 @@ apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.alexeymerov.unsplashviewer"
         minSdkVersion 21
-        targetSdkVersion 27
-        versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 28
+        versionCode 2
+        versionName "1.1"
         vectorDrawables.useSupportLibrary = true
         resConfigs "en", "ru"
     }
+
     buildTypes {
         release {
-            //            debuggable false
-//            jniDebuggable false
-//            renderscriptDebuggable false
-//            pseudoLocalesEnabled false
-//            zipAlignEnabled true
-            //            ndk {
+            debuggable false
+            jniDebuggable false
+            renderscriptDebuggable false
+            pseudoLocalesEnabled false
+//            ndk {
 //                abiFilters "arm64-v8a", "armeabi-v7a"
 //            }
 
-
-            minifyEnabled false
-            shrinkResources false
+            minifyEnabled true
+            shrinkResources true
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
             buildOutputApkName(applicationVariants)
         }
         debug {
             minifyEnabled false
             shrinkResources false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
             buildOutputApkName(applicationVariants)
         }
     }
