@@ -1,14 +1,17 @@
 package com.alexeymerov.unsplashviewer.domain
 
 import android.app.Application
-import android.arch.lifecycle.AndroidViewModel
-import android.support.annotation.StringRes
-import com.alexeymerov.unsplashviewer.App
+import androidx.lifecycle.AndroidViewModel
+import com.alexeymerov.unsplashviewer.data.repository.BaseRepository
 
-abstract class BaseViewModel(application: Application) : AndroidViewModel(application) {
+abstract class BaseViewModel(application: Application,
+                             private vararg val repositories: BaseRepository? = arrayOfNulls(0))
+    : AndroidViewModel(application) {
 
-    fun getContext() = getApplication<App>()
-
-    fun getString(@StringRes id: Int): String = getContext().getString(id)
-
+    override fun onCleared() {
+        for (repo in repositories) {
+            repo?.clean()
+        }
+        super.onCleared()
+    }
 }
