@@ -1,6 +1,6 @@
 package com.alexeymerov.unsplashviewer.utils;
 
-import android.support.v4.content.FileProvider;
+import androidx.core.content.FileProvider;
 
 public class GenericFileProvider extends FileProvider {
 }
