@@ -1,6 +1,6 @@
 package com.alexeymerov.unsplashviewer.data.server.pojo.response
 
-import com.alexeymerov.unsplashviewer.data.database.entity.ImageEntity
+import com.alexeymerov.unsplashviewer.data.entity.ImageEntity
 import com.squareup.moshi.Json
 
 data class SearchResponse(
