@@ -3,5 +3,5 @@ package com.alexeymerov.unsplashviewer.utils
 import com.bumptech.glide.annotation.GlideModule
 import com.bumptech.glide.module.AppGlideModule
 
-@GlideModule()
+@GlideModule
 class AppGlideModule : AppGlideModule()
